+++ src/Components/ServiceScene/ServiceScene.tsx
@@ -43,7 +43,7 @@ import { sortLabelsByCardinality } from 'services/filters';
 import { SERVICE_NAME } from 'Components/ServiceSelectionScene/ServiceSelectionScene';
 import { getMetadataService } from '../../services/metadata';
 import { navigateToDrilldownPage, navigateToIndex } from '../../services/navigate';
-import { areArraysEqual } from '../../services/state';
+import { areArraysEqual } from '../../services/comparison';
 
 export interface LokiPattern {
   pattern: string;
